Revert "Convert RetrofitResponseTypeKeepProcessorTest to parametermized tests"
