--- retrofit-response-type-keeper/src/test/resources/kotlin-suspend/Service.pro
@@ -1,3 +1,0 @@
-# test.Service
--keep,allowoptimization,allowshrinking,allowobfuscation class java.lang.Object
--keep,allowoptimization,allowshrinking,allowobfuscation class test.Body
